Rewrites this gem based on simple-service.
I figured out that simple-cli many concerns with other packages (postjob and simple-httpd). All of these packages contain code which connects "native" ruby code, in the form of instance methods in a module (a "service") and connects these with various "request" sources (for example a HTTP request, or a job task, or a CLI invocation.)

In light of this this commit extracts everything service-related from this gem, moves this into "simple-service", and adjusts calls to this, now external, code. As a result this commit is more like a complete rerite of simple-cli, which collected some wrinkles over the course of the 0.2.* releases.